moved ViewModelPreviewData into its own file,
This is more of a test for GitHub token that was imported into xcode.
